update FindbyID in models

+++ models/agent.go
@@ -57,7 +57,7 @@ func (model *Agent) Delete() error {
 }
 
 // FindbyID find agent with id
-func (model *Agent) FindbyID(id int) error {
+func (model *Agent) FindbyID(id uint64) error {
 	err := basemodel.FindbyID(&model, id)
 	return err
 }
@@ -79,7 +79,7 @@ func (model *Agent) PagedFilterSearch(page int, rows int, order []string, sort [
 // checkBorrowerID search using filter and return last
 func (model *Agent) CheckBorrowerOwnedByAgent(borrowerID uint64) bool {
 	borrowerModel := Borrower{}
-	err := borrowerModel.FindbyID(int(borrowerID))
+	err := borrowerModel.FindbyID(borrowerID)
 	if err != nil {
 		return false
 	}
