+++ models/borrower.go
@@ -106,7 +106,7 @@ func (b *Borrower) Unsuspend() error {
 	return err
 }
 
-func (b *Borrower) FindbyID(id int) error {
+func (b *Borrower) FindbyID(id uint64) error {
 	err := basemodel.FindbyID(&b, id)
 	return err
 }
