+++ models/user.go
@@ -54,7 +54,7 @@ func (model *User) Delete() error {
 }
 
 // FindbyID find User with id
-func (model *User) FindbyID(id int) error {
+func (model *User) FindbyID(id uint64) error {
 	err := basemodel.FindbyID(&model, id)
 	return err
 }
